--- app/server/config.env
@@ -1,5 +1,0 @@
-DBHost=localhost
-DBPort=5432
-DBUser=postgres
-DBPassword=secret
-DBName=rollertempo
