--- app/server/controllers/utils.go
@@ -1,9 +1,0 @@
-package controllers
-
-const (
-	ErrorMessage400            string = "El identificador ingresado no es válido"
-	ErrorMessage404            string = "Error al obtener información del identificador especificado"
-	ErrorMessage500            string = "Hubo un error en el procesamiento interno"
-	ErrorMessageRegisteredUser string = "Ya existe un registro con el identificador especificado"
-	OkMessageRegistry          string = "Registro exitoso"
-)
