+++ app/server/model/reward.go
@@ -1,0 +1,11 @@
+package models
+
+import "gorm.io/gorm"
+
+type Reward struct {
+	gorm.Model
+
+	name        string
+	description string
+	price       int
+}
