+++ app/server/utils/migrations.go
@@ -1,0 +1,1 @@
+package utils
